pyproject.toml
[project]
name = "innovathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp-common-utils",
    "modelcontextprotocol>=0.1.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
]

[tool.uv.sources]
mcp-common-utils = { path = "common", editable = true }

[tool.uv.scripts]
start-datetime = "python server_datetime/main.py"
